// ets2panda/bindings/src/generated/Es2pandaEnums.ts
export enum Es2pandaContextState {
  ES2PANDA_STATE_NEW = 0,
  ES2PANDA_STATE_PARSED = 1,
  ES2PANDA_STATE_SCOPE_INITED = 2,
  ES2PANDA_STATE_BOUND = 3,
  ES2PANDA_STATE_CHECKED = 4,
  ES2PANDA_STATE_LOWERED = 5,
  ES2PANDA_STATE_ASM_GENERATED = 6,
  ES2PANDA_STATE_BIN_GENERATED = 7,
  ES2PANDA_STATE_ERROR = 8
}